requirements.txt
absl-py==2.1.0
altair==4.2.2
astunparse==1.6.3
attrs==24.2.0
audio-recorder-streamlit==0.0.10
bcrypt==4.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.5.0
captcha==0.6.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
Cython==3.0.11
deepface==0.0.93
dnspython==2.6.1
entrypoints==0.4
extra-streamlit-components==0.1.71
filelock==3.16.1
fire==0.6.0
Flask==3.0.3
Flask-Cors==5.0.0
flatbuffers==24.3.25
gast==0.6.0
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.66.1
gunicorn==23.0.0
h5py==3.12.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==3.5.0
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.4.1
mtcnn==0.1.1
namex==0.0.8
narwhals==1.8.3
numpy==1.26.4
opencv-python==4.10.0.84
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pocketsphinx==5.0.3
protobuf==4.25.5
pyarrow==17.0.0
pycparser==2.22
pycryptodome==3.20.0
pydeck==0.9.1
pydub==0.25.1
Pygments==2.18.0
PyJWT==2.9.0
pymongo==4.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
retina-face==0.0.17
rich==13.8.1
rpds-py==0.20.0
setuptools==75.1.0
six==1.16.0
smmap==5.0.1
soundfile==0.12.1
soupsieve==2.6
speechbrain==1.0.1
SpeechRecognition==3.10.4
streamlit==1.38.0
streamlit-authenticator==0.3.3
streamlit-mic-recorder==0.0.8
tenacity==8.5.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io==0.37.1
termcolor==2.4.0
tf_keras==2.17.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==4.0.2
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
plotly
